beeware-venv/Scripts/atrportable/android/gradle/ATRportable/app/requirements.txt
numpy
opencv-python
toga-android>=0.3.0.dev38
numpy
opencv-python
